requirements.txt
fastapi
uvicorn
motor
pydantic-settings
fastapi-mail
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
dnspython
email-validator
apscheduler
